requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
decorator==5.1.1
decorators==2.0.7
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
executing==2.0.1
export==0.2.1
gevent==23.9.1
greenlet==3.0.1
idna==3.4
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==5.3.3
logging==0.4.9.6
matplotlib-inline==0.1.6
openapi==1.1.0
packaging==23.2
parso==0.8.3
prompt-toolkit==3.0.40
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.9
zope.event==5.0
zope.interface==6.1
